Configurable IPGeo service (#206)
* makes IPGeo service configurable

* adds translation of json to support different service responses
